update design-tokens and isolate jest config
