--- jest.config.ts
@@ -1,17 +1,0 @@
-import { JestConfigWithTsJest } from "ts-jest"
-
-const jestConfig: JestConfigWithTsJest = {
-  preset: "ts-jest",
-  testEnvironment: "jsdom",
-  testMatch: ["**/*.jest.spec.ts?(x)"],
-  setupFilesAfterEnv: ["jest-canvas-mock", "<rootDir>/jest.setup.ts"],
-  moduleNameMapper: {
-    "\\.(jpe?g|png|webm|mp4)$": "jest-static-stubs/$1",
-    "\\.s?css$": "identity-obj-proxy",
-  },
-  transformIgnorePatterns: ["[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"],
-}
-
-export default jestConfig
-
-process.env.TZ = "UTC"
