--- packages/design-tokens/jest.config.ts
@@ -1,2 +1,0 @@
-import sharedConfig from "../../jest.config"
-export default sharedConfig
