--- packages/design-tokens/jest.setup.ts
@@ -1,1 +1,0 @@
-import "../../jest.setup"
