+++ packages/design-tokens/src/__tests__/sass-tokens.spec.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/consistent-test-it */
 import fs from "fs"
 import path from "path"
 import PostCSS, { Declaration } from "postcss"
@@ -28,7 +27,7 @@ ${typography.toString()}
     `
   )
   describe("everything", () => {
-    test("no tokens start with kz", () => {
+    it("no tokens start with kz", () => {
       const allTokensStartingWithKz = everything.nodes
         .filter(
           (n): n is Declaration => n.type === "decl" && n.prop.startsWith("$kz")
@@ -39,7 +38,7 @@ ${typography.toString()}
   })
 
   describe("colors", () => {
-    test("new un-prefixed tokens only contain heart color names", () => {
+    it("new un-prefixed tokens only contain heart color names", () => {
       const newColors = colors.nodes.filter(
         (n): n is Declaration =>
           n.type === "decl" && n.prop.startsWith("$color")
@@ -56,7 +55,7 @@ ${typography.toString()}
   })
 
   describe("layout", () => {
-    test("new un-prefixed layout tokens are not CSS variables", () => {
+    it("new un-prefixed layout tokens are not CSS variables", () => {
       expect(
         layout.nodes.filter(
           n =>
