+++ packages/design-tokens/tsconfig.json
@@ -7,7 +7,8 @@
       "~design-tokens/*": ["../design-tokens/src/*"],
       "~components/*": ["../components/src/*"],
       "~tailwind/*": ["./src/*"]
-    }
+    },
+    "types": ["@testing-library/jest-dom", "vitest/globals"]
   },
   "include": ["src/**/*", "_docs/**/*"],
   "files": ["./types.d.ts"]
