+++ tsconfig.json
@@ -19,8 +19,7 @@
       "~tailwind/*": ["./packages/tailwind/src/*"],
       "~tests": ["./packages/components/__tests__/index"],
       "~tests/*": ["./packages/components/__tests__/*"]
-    },
-    "types": ["@testing-library/jest-dom"]
+    }
   },
   "exclude": ["**/node_modules"]
 }
